backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.3.0
openai==1.12.0
python-jose==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==24.1
django-storages==1.14.2
boto3==1.34.34
django-allauth==0.58.2
django-rest-auth==5.0.0
django-celery==5.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.2.0
django-cleanup==7.0.0
django-environ==0.11.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3 